Implemented opening and closing the index file
I had to make some changes to the structure of the JSON file so I have updated the Wiki to reflect. You will want to copy the new structure from the Wiki and put in your local copy of the index. I also removed the AbstractTableClass as a parent of IndexModel as this was causing issues when saving to disk. We may want to implement a separate class for our own TableModel if it becomes necessary.

Fixes #2, Fixes #3, Fixes #8